+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Change log
 
 ## in dev
+- `luid-user-picker` - allows to search a user by his employee number
 - `luid-table-grid`: adds support for columns whose content must preserve line-breaks
+
 ### Bug fixes
 
 ## 3.1.16 - [release](https://github.com/LuccaSA/lucca-ui/releases/tag/3.1.16)